added risingwave chisquared udf

+++ container/risingwave_udf_chisquared/chisquared.py
@@ -13,7 +13,7 @@ def chisquared_numeric(data, data_mean, data_std):
     if data_array.size == 0 or data_mean == None or data_std == None or data_mean == 0 or data_std == 0:
         return Decimal(0.0)
     else:
-        return (((data_array - data_mean)/data_std)**2).sum()/data_array.size
+        return Decimal((((data_array - data_mean)/data_std)**2).sum()/data_array.size)
 
 # Define a table function
 @udf(input_types=['real[]', 'double precision', 'double precision'], result_type='numeric')
@@ -22,7 +22,7 @@ def chisquared_double(data, data_mean, data_std):
     if data_array.size == 0 or data_mean == None or data_std == None or data_mean == 0 or data_std == 0:
         return Decimal(0.0)
     else:
-        return (((data_array - data_mean)/data_std)**2).sum()/data_array.size
+        return Decimal((((data_array - data_mean)/data_std)**2).sum()/data_array.size)
